drlhp/laryn-requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
decorator==4.4.2
idna==3.4
imageio==2.31.6
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.23.5
opencv-python==4.8.1.78
Pillow==10.0.1
proglog==0.1.10
pygame==2.5.2
pyglet==2.0.9
requests==2.31.0
scikit-video==1.1.11
scipy==1.11.3
tqdm==4.66.1
urllib3==2.0.7